Docs (README.md): update documentation for conditional logic
Update documentation to include how to apply conditional logic to a
field when it is inside of a sub-field.
How to do this is no apparent to me in the documentation. There is a
closed issue that explains it, but I believe this is a common enough use
case to warrant being in the documentation.